Fix initializing state logic
This function contains logic to look for a phase field or an "active"
field on the spec, if the resource has one and to use those
values to determine the resource's state.

But, if the resource was older than five seconds and had no conditions,
that logic was not getting hit and the resource was just defaulting
to "active".

This change let's us check the phase and spec.active before falling
back to active or initializing.