sort env key
Problems:
When we created workloads in rancher api, the environment variable
field is converted without a specific order.(we simplely loop in the
map). This will cause K8s workload spec to change constantly if we do a
update, and cause the pod getting recreated.
Solutions:
Sort the map and make sure they always follow the same order in k8s
workload spec.