Check data is nil before executing mapping in os_info
This os_info mapper is used for v1.NodeStatus but
there is some case that mapper is evaluated when
v1.NodeStatus is nil.  For example, when the user try
to create v3.Node by calling API(POST /v3/nodes).

So we should check data is nil or not before executing
mapper in order to prevent from trying to use data(nil)
as map inside OSInfo.FromInternal