Allow a load-balancer address for k8s API
This optional string is to replace the hostname used by other nodes and
configurations when attempting contacting the k8s API service.

It is expected to be the hostname of a layer-4 (TCP) load-balancer which
will be externally configured by the organisation to direct traffic to
an appropriate selection of the cluster's ControlPlane nodes on the same
port as it arrived; i.e. TCP/6443 for kube-apiserver.

Any TLS configuration and healthchecks on the load-balancer host are
also expected to be managed by the organisation.